requirements.txt
PyQt5>=5.15.0
requests>=2.25.0
python-dotenv>=0.19.0
